% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdgp654.R
\docType{data}
\name{bdgp654}
\alias{bdgp654}
\title{Fruitfly annotation data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 28759 rows and 9 columns.
}
\source{
\url{http://ensembl.org/drosophila_melanogaster}
}
\usage{
bdgp654
}
\description{
Fruitfly (\emph{Drosophila melanogaster}) annotations based on
genome assembly BDGP654 from Ensembl.
}
\details{
Variables:
\itemize{
\item ensgene
\item entrez
\item symbol
\item chr
\item start
\item end
\item strand
\item biotype
\item description
}
}
\examples{
head(bdgp654)
}
\keyword{datasets}
