% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdgp6_tx2gene.R
\docType{data}
\name{bdgp6_tx2gene}
\alias{bdgp6_tx2gene}
\title{Fruitfly transcripts to genes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 41209 rows and 2 columns.
}
\source{
\url{http://ensembl.org/drosophila_melanogaster}
}
\usage{
bdgp6_tx2gene
}
\description{
Lookup table for converting Fruitfly (\emph{Drosophila melanogaster})
Ensembl transcript IDs to gene IDs based on genome assembly
BDGP6 from Ensembl.
}
\details{
Variables:
\itemize{
\item enstxp
\item ensgene
}
}
\examples{
head(bdgp6_tx2gene)
}
\keyword{datasets}
