% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfamiliaris.R
\docType{data}
\name{cfamiliaris}
\alias{cfamiliaris}
\title{Dog gene annotation table}
\format{
A data frame with gene annotation columns.
}
\source{
Ensembl

\url{http://ensembl.org/canis_lupus familiaris}
}
\usage{
cfamiliaris
}
\description{
Gene annotation table for Canis lupus familiaris (dog) from Ensembl.
}
\details{
Variables:
\itemize{
\item ensgene
\item entrez
\item symbol
\item chr
\item start
\item end
\item strand
\item biotype
\item description
}
}
\examples{
head(cfamiliaris)
}
\keyword{datasets}
