% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfamiliaris_tx2gene.R
\docType{data}
\name{cfamiliaris_tx2gene}
\alias{cfamiliaris_tx2gene}
\title{Dog transcript-to-gene table}
\format{
A data frame with transcript and gene columns.
}
\source{
Ensembl
}
\usage{
cfamiliaris_tx2gene
}
\description{
Transcript-to-gene mapping for Canis lupus familiaris (dog) from Ensembl.
}
\keyword{datasets}
