% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drerio.R
\docType{data}
\name{drerio}
\alias{drerio}
\title{Zebrafish gene annotation table}
\format{
A data frame with gene annotation columns.
}
\source{
Ensembl

\url{http://ensembl.org/danio_rerio}
}
\usage{
drerio
}
\description{
Gene annotation table for Danio rerio (zebrafish) from Ensembl.
}
\details{
Variables:
\itemize{
\item ensgene
\item entrez
\item symbol
\item chr
\item start
\item end
\item strand
\item biotype
\item description
}
}
\examples{
head(drerio)
}
\keyword{datasets}
