% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drerio_tx2gene.R
\docType{data}
\name{drerio_tx2gene}
\alias{drerio_tx2gene}
\title{Zebrafish transcript-to-gene table}
\format{
A data frame with transcript and gene columns.
}
\source{
Ensembl
}
\usage{
drerio_tx2gene
}
\description{
Transcript-to-gene mapping for Danio rerio (zebrafish) from Ensembl.
}
\keyword{datasets}
