% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_version.R
\docType{data}
\name{ensembl_version}
\alias{ensembl_version}
\title{Ensembl version used for build}
\format{
An object of class \code{character} of length 1.
}
\source{
\url{http://www.ensembl.org/}
}
\usage{
ensembl_version
}
\description{
Ensembl version used for build
}
\examples{
ensembl_version

}
\keyword{datasets}
