% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galgal4.R
\docType{data}
\name{galgal4}
\alias{galgal4}
\title{Chicken annotation Galgal4}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 17680 rows and 9 columns.}
\source{
\url{http://useast.ensembl.org/Gallus_gallus}
}
\usage{
galgal4
}
\description{
Chicken annotation Galgal4 from Ensembl release 82.
}
\examples{
head(galgal4)

}
\keyword{datasets}

