% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galgal5.R
\docType{data}
\name{galgal5}
\alias{galgal5}
\title{Chicken annotation Galgal5}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 25677 rows and 9 columns.}
\source{
\url{http://useast.ensembl.org/Gallus_gallus}
}
\usage{
galgal5
}
\description{
Chicken annotation Galgal5 from Ensembl.
}
\examples{
head(galgal5)

}
\keyword{datasets}

