% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galgal5_tx2gene.R
\docType{data}
\name{galgal5_tx2gene}
\alias{galgal5_tx2gene}
\title{Chicken (galgal5) transcripts to genes}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 38118 rows and 2 columns.}
\source{
\url{http://www.ensembl.org/}
}
\usage{
galgal5_tx2gene
}
\description{
Chicken (galgal5) Ensembl transcript IDs to gene IDs.
}
\examples{
head(galgal5_tx2gene)

}
\keyword{datasets}

