% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grcg7b.R
\docType{data}
\name{grcg7b}
\alias{grcg7b}
\title{Chicken annotation data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 34332 rows and 9 columns.
}
\source{
\url{http://ensembl.org/gallus_gallus}
}
\usage{
grcg7b
}
\description{
Chicken (\emph{Gallus gallus}) annotations based on
genome assembly GRCG7B from Ensembl.
}
\details{
Variables:
\itemize{
\item ensgene
\item entrez
\item symbol
\item chr
\item start
\item end
\item strand
\item biotype
\item description
}
}
\examples{
head(grcg7b)
}
\keyword{datasets}
