% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grcg7b_tx2gene.R
\docType{data}
\name{grcg7b_tx2gene}
\alias{grcg7b_tx2gene}
\title{Chicken transcripts to genes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 72689 rows and 2 columns.
}
\source{
\url{http://ensembl.org/gallus_gallus}
}
\usage{
grcg7b_tx2gene
}
\description{
Lookup table for converting Chicken (\emph{Gallus gallus})
Ensembl transcript IDs to gene IDs based on genome assembly
GRCG7B from Ensembl.
}
\details{
Variables:
\itemize{
\item enstxp
\item ensgene
}
}
\examples{
head(grcg7b_tx2gene)
}
\keyword{datasets}
