% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grch37.R
\docType{data}
\name{grch37}
\alias{grch37}
\title{Human annotation GRCh37}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 67416 rows and 9 columns.}
\source{
\url{http://grch37.ensembl.org/}
}
\usage{
grch37
}
\description{
Human annotation GRCh37 from Ensembl release 75.
}
\examples{
head(grch37)

}
\keyword{datasets}

