% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grch37_gt.R
\docType{data}
\name{grch37_gt}
\alias{grch37_gt}
\title{Human GRCh37 genes-to-transcripts}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 215647 rows and 2 columns.}
\source{
\url{http://grch37.ensembl.org/}
}
\usage{
grch37_gt
}
\description{
Human GRCh37 Ensembl gene IDs to transcript IDs from Ensembl release 75.
}
\examples{
head(grch37_gt)

}
\keyword{datasets}

