% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grch37_tx2gene.R
\docType{data}
\name{grch37_tx2gene}
\alias{grch37_tx2gene}
\title{Human GRCh37 transcripts to genes}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 215170 rows and 2 columns.}
\source{
\url{http://www.ensembl.org/}
}
\usage{
grch37_tx2gene
}
\description{
Human GRCh37  Ensembl transcript IDs to gene IDs.
}
\examples{
head(grch37_tx2gene)

}
\keyword{datasets}

