% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grch38_gt.R
\docType{data}
\name{grch38_gt}
\alias{grch38_gt}
\title{Human GRCh38.p3 genes-to-transcripts}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 216133 rows and 2 columns.}
\source{
\url{http://useast.ensembl.org/Homo_sapiens}
}
\usage{
grch38_gt
}
\description{
Human GRCh38.p3 Ensembl gene IDs to transcript IDs from Ensembl release 82.
}
\examples{
head(grch38_gt)

}
\keyword{datasets}

