% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grcm38.R
\docType{data}
\name{grcm38}
\alias{grcm38}
\title{Mouse annotation data}
\format{An object of class \code{data.frame} with 53001 rows and 9 columns.}
\source{
\url{http://ensembl.org/mus_musculus}
}
\usage{
grcm38
}
\description{
Mouse (\emph{Mus musculus}) annotations based on genome assembly GRCM38 from Ensembl.
}
\details{
Variables:

\itemize{
\item ensgene
\item entrez
\item symbol
\item chr
\item start
\item end
\item strand
\item biotype
\item description
}
}
\examples{
head(grcm38)

}
\keyword{datasets}
