% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grcm38_gt.R
\docType{data}
\name{grcm38_gt}
\alias{grcm38_gt}
\title{Mouse GRCm38.p4 genes-to-transcripts}
\format{An object of class \code{data.frame} with 112517 rows and 2 columns.}
\source{
\url{http://useast.ensembl.org/Mus_musculus}
}
\usage{
grcm38_gt
}
\description{
Mouse GRCm38.p4 Ensembl gene IDs to transcript IDs from Ensembl release 82.
}
\keyword{datasets}

