% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grcm38_tg.R
\docType{data}
\name{grcm38_tg}
\alias{grcm38_tg}
\title{Mouse GRCm38.p4 genes-to-transcripts}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 112517 rows and 2 columns.}
\source{
\url{http://useast.ensembl.org/Mus_musculus}
}
\usage{
grcm38_tg
}
\description{
Mouse GRCm38.p4 Ensembl transcript IDs to gene IDs from Ensembl release 82.
}
\examples{
head(grcm38_tg)

}
\keyword{datasets}

