% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grcm38_tx2gene.R
\docType{data}
\name{grcm38_tx2gene}
\alias{grcm38_tx2gene}
\title{Mouse GRCm38 transcripts to genes}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 124168 rows and 2 columns.}
\source{
\url{http://www.ensembl.org/}
}
\usage{
grcm38_tx2gene
}
\description{
Mouse GRCm38 Ensembl transcript IDs to gene IDs.
}
\examples{
head(grcm38_tx2gene)

}
\keyword{datasets}

