% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grcm39.R
\docType{data}
\name{grcm39}
\alias{grcm39}
\title{Mouse annotation data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 78718 rows and 9 columns.
}
\source{
\url{http://ensembl.org/mus_musculus}
}
\usage{
grcm39
}
\description{
Mouse (\emph{Mus musculus}) annotations based on
genome assembly GRCM39 from Ensembl.
}
\details{
Variables:
\itemize{
\item ensgene
\item entrez
\item symbol
\item chr
\item start
\item end
\item strand
\item biotype
\item description
}
}
\examples{
head(grcm39)
}
\keyword{datasets}
