% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grcm39_tx2gene.R
\docType{data}
\name{grcm39_tx2gene}
\alias{grcm39_tx2gene}
\title{Mouse transcripts to genes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 481956 rows and 2 columns.
}
\source{
\url{http://ensembl.org/mus_musculus}
}
\usage{
grcm39_tx2gene
}
\description{
Lookup table for converting Mouse (\emph{Mus musculus})
Ensembl transcript IDs to gene IDs based on genome assembly
GRCM39 from Ensembl.
}
\details{
Variables:
\itemize{
\item enstxp
\item ensgene
}
}
\examples{
head(grcm39_tx2gene)
}
\keyword{datasets}
