% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grcr8.R
\docType{data}
\name{grcr8}
\alias{grcr8}
\title{Rat annotation data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 57760 rows and 9 columns.
}
\source{
\url{http://ensembl.org/rattus_norvegicus}
}
\usage{
grcr8
}
\description{
Rat (\emph{Rattus norvegicus}) annotations based on
genome assembly GRCR8 from Ensembl.
}
\details{
Variables:
\itemize{
\item ensgene
\item entrez
\item symbol
\item chr
\item start
\item end
\item strand
\item biotype
\item description
}
}
\examples{
head(grcr8)
}
\keyword{datasets}
