% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grcz11.R
\docType{data}
\name{grcz11}
\alias{grcz11}
\title{Zebrafish annotation data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 92137 rows and 9 columns.
}
\source{
\url{http://ensembl.org/danio_rerio}
}
\usage{
grcz11
}
\description{
Zebrafish (\emph{Danio rerio}) annotations based on
genome assembly GRCZ11 from Ensembl.
}
\details{
Variables:
\itemize{
\item ensgene
\item entrez
\item symbol
\item chr
\item start
\item end
\item strand
\item biotype
\item description
}
}
\examples{
head(grcz11)
}
\keyword{datasets}
