% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grcz11_tx2gene.R
\docType{data}
\name{grcz11_tx2gene}
\alias{grcz11_tx2gene}
\title{Zebrafish transcripts to genes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 65905 rows and 2 columns.
}
\source{
\url{http://ensembl.org/danio_rerio}
}
\usage{
grcz11_tx2gene
}
\description{
Lookup table for converting Zebrafish (\emph{Danio rerio})
Ensembl transcript IDs to gene IDs based on genome assembly
GRCZ11 from Ensembl.
}
\details{
Variables:
\itemize{
\item enstxp
\item ensgene
}
}
\examples{
head(grcz11_tx2gene)
}
\keyword{datasets}
