% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmul10.R
\docType{data}
\name{mmul10}
\alias{mmul10}
\title{Macaque annotation data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 37169 rows and 9 columns.
}
\source{
\url{http://ensembl.org/macaca_mulatta}
}
\usage{
mmul10
}
\description{
Macaque (\emph{Macaca mulatta}) annotations based on
genome assembly MMUL10 from Ensembl.
}
\details{
Variables:
\itemize{
\item ensgene
\item entrez
\item symbol
\item chr
\item start
\item end
\item strand
\item biotype
\item description
}
}
\examples{
head(mmul10)
}
\keyword{datasets}
