% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmul801.R
\docType{data}
\name{mmul801}
\alias{mmul801}
\title{Macaque annotation data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 32820 rows and 9 columns.}
\source{
\url{http://ensembl.org/macaca_mulatta}
}
\usage{
mmul801
}
\description{
Macaque (\emph{Macaca mulatta}) annotations based on
genome assembly MMUL801 from Ensembl.
}
\details{
Variables:
\itemize{
\item ensgene
\item entrez
\item symbol
\item chr
\item start
\item end
\item strand
\item biotype
\item description
}
}
\examples{
head(mmul801)
}
\keyword{datasets}
