% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnor6_tx2gene.R
\docType{data}
\name{rnor6_tx2gene}
\alias{rnor6_tx2gene}
\title{Rat rnor6 transcripts to genes}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 40459 rows and 2 columns.}
\source{
\url{http://www.ensembl.org/}
}
\usage{
rnor6_tx2gene
}
\description{
Rat rnor6 Ensembl transcript IDs to gene IDs.
}
\examples{
head(rnor6_tx2gene)

}
\keyword{datasets}

