% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roscfam1.R
\docType{data}
\name{roscfam1}
\alias{roscfam1}
\title{Dog annotation data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 34012 rows and 9 columns.
}
\source{
\url{http://ensembl.org/canis_lupus familiaris}
}
\usage{
roscfam1
}
\description{
Dog (\emph{Canis lupus familiaris}) annotations based on
genome assembly ROSCFAM1 from Ensembl.
}
\details{
Variables:
\itemize{
\item ensgene
\item entrez
\item symbol
\item chr
\item start
\item end
\item strand
\item biotype
\item description
}
}
\examples{
head(roscfam1)
}
\keyword{datasets}
