% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roscfam1_tx2gene.R
\docType{data}
\name{roscfam1_tx2gene}
\alias{roscfam1_tx2gene}
\title{Dog transcripts to genes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 55335 rows and 2 columns.
}
\source{
\url{http://ensembl.org/canis_lupus familiaris}
}
\usage{
roscfam1_tx2gene
}
\description{
Lookup table for converting Dog (\emph{Canis lupus familiaris})
Ensembl transcript IDs to gene IDs based on genome assembly
ROSCFAM1 from Ensembl.
}
\details{
Variables:
\itemize{
\item enstxp
\item ensgene
}
}
\examples{
head(roscfam1_tx2gene)
}
\keyword{datasets}
