% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sscrofa.R
\docType{data}
\name{sscrofa}
\alias{sscrofa}
\title{Pig annotation data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 35819 rows and 9 columns.
}
\source{
\url{http://ensembl.org/sus_scrofa}
}
\usage{
sscrofa
}
\description{
Pig (\emph{Sus scrofa}) annotations based on
genome assembly SSCROFA from Ensembl.
}
\details{
Variables:
\itemize{
\item ensgene
\item entrez
\item symbol
\item chr
\item start
\item end
\item strand
\item biotype
\item description
}
}
\examples{
head(sscrofa)
}
\keyword{datasets}
