% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sscrofa_tx2gene.R
\docType{data}
\name{sscrofa_tx2gene}
\alias{sscrofa_tx2gene}
\title{Pig transcripts to genes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 60440 rows and 2 columns.
}
\source{
\url{http://ensembl.org/sus_scrofa}
}
\usage{
sscrofa_tx2gene
}
\description{
Lookup table for converting Pig (\emph{Sus scrofa})
Ensembl transcript IDs to gene IDs based on genome assembly
SSCROFA from Ensembl.
}
\details{
Variables:
\itemize{
\item enstxp
\item ensgene
}
}
\examples{
head(sscrofa_tx2gene)
}
\keyword{datasets}
