% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbcel235.R
\docType{data}
\name{wbcel235}
\alias{wbcel235}
\title{Roundworm annotation data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 46926 rows and 9 columns.
}
\source{
\url{http://ensembl.org/caenorhabditis_elegans}
}
\usage{
wbcel235
}
\description{
Roundworm (\emph{Caenorhabditis elegans}) annotations based on
genome assembly WBCEL235 from Ensembl.
}
\details{
Variables:
\itemize{
\item ensgene
\item entrez
\item symbol
\item chr
\item start
\item end
\item strand
\item biotype
\item description
}
}
\examples{
head(wbcel235)
}
\keyword{datasets}
