% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbcel235_tx2gene.R
\docType{data}
\name{wbcel235_tx2gene}
\alias{wbcel235_tx2gene}
\title{C. elegans wbcel235 transcripts to genes}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 58941 rows and 2 columns.}
\source{
\url{http://www.ensembl.org/}
}
\usage{
wbcel235_tx2gene
}
\description{
C. elegans wbcel235 Ensembl transcript IDs to gene IDs.
}
\examples{
head(wbcel235_tx2gene)

}
\keyword{datasets}

