% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotables.R
\docType{package}
\name{annotables-package}
\alias{annotables}
\alias{annotables-package}
\title{annotables}
\description{
R data package for annotating/converting Gene IDs.
}
\author{
\strong{Maintainer}: Stephen Turner \email{vustephen@gmail.com}

Authors:
\itemize{
  \item Michael J. Steinbaugh \email{mike@steinbaugh.com}
}

}
