% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotables.R
\docType{package}
\name{annotables-package}
\alias{annotables}
\alias{annotables-package}
\title{annotables}
\description{
Provides tables for converting and annotating Ensembl Gene IDs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stephenturner/annotables}
  \item Report bugs at \url{https://github.com/stephenturner/annotables/issues}
}

}
\author{
\strong{Maintainer}: Stephen Turner \email{vustephen@gmail.com}

Other contributors:
\itemize{
  \item Michael J. Steinbaugh \email{mike@steinbaugh.com} [contributor]
  \item Aaron Wolen \email{aaron@wolen.com} [contributor]
}

}
