% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdgp6.R
\docType{data}
\name{bdgp6}
\alias{bdgp6}
\title{Fly annotation BDGP6}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 19639 rows and 9 columns.}
\source{
\url{http://useast.ensembl.org/Drosophila_melanogaster}
}
\usage{
bdgp6
}
\description{
Fly annotation BDGP6 from Ensembl release 82.
}
\keyword{datasets}

