% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdgp6_tx2gene.R
\docType{data}
\name{bdgp6_tx2gene}
\alias{bdgp6_tx2gene}
\title{Fly BDGP6 transcripts to genes}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 34749 rows and 2 columns.}
\source{
\url{http://www.ensembl.org/}
}
\usage{
bdgp6_tx2gene
}
\description{
Fly BDGP6 Ensembl transcript IDs to gene IDs.
}
\examples{
head(bdgp6_tx2gene)

}
\keyword{datasets}

