% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galgal5.R
\docType{data}
\name{galgal5}
\alias{galgal5}
\title{Chicken annotation data}
\format{An object of class \code{data.frame} with 25677 rows and 9 columns.}
\source{
\url{http://ensembl.org/gallus_gallus}
}
\usage{
galgal5
}
\description{
Chicken (\emph{Gallus gallus}) annotations based on genome assembly GALGAL5 from Ensembl.
}
\details{
Variables:

\itemize{
\item ensgene
\item entrez
\item symbol
\item chr
\item start
\item end
\item strand
\item biotype
\item description
}
}
\examples{
head(galgal5)

}
\keyword{datasets}
