% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grch37.R
\docType{data}
\name{grch37}
\alias{grch37}
\title{Human annotation data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 66978 rows and 9 columns.}
\source{
\url{http://ensembl.org/homo_sapiens}
}
\usage{
grch37
}
\description{
Human (\emph{Homo sapiens}) annotations based on
genome assembly GRCH37 from Ensembl.
}
\details{
Variables:
\itemize{
\item ensgene
\item entrez
\item symbol
\item chr
\item start
\item end
\item strand
\item biotype
\item description
}
}
\examples{
head(grch37)
}
\keyword{datasets}
