% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grch38.R
\docType{data}
\name{grch38}
\alias{grch38}
\title{Human annotation GRCh38.p3}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 66531 rows and 9 columns.}
\source{
\url{http://useast.ensembl.org/Homo_sapiens}
}
\usage{
grch38
}
\description{
Human annotation GRCh38.p3 from Ensembl release 82.
}
\keyword{datasets}

