% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grch38_tx2gene.R
\docType{data}
\name{grch38_tx2gene}
\alias{grch38_tx2gene}
\title{Human GRCh38 transcripts to genes}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 215929 rows and 2 columns.}
\source{
\url{http://www.ensembl.org/}
}
\usage{
grch38_tx2gene
}
\description{
Human GRCh38 Ensembl transcript IDs to gene IDs.
}
\examples{
head(grch38_tx2gene)

}
\keyword{datasets}

