% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grch38_tx2gene.R
\docType{data}
\name{grch38_tx2gene}
\alias{grch38_tx2gene}
\title{Human transcripts to genes}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 219288 rows and 2 columns.}
\source{
\url{http://ensembl.org/homo_sapiens}
}
\usage{
grch38_tx2gene
}
\description{
Lookup table for converting Human (\emph{Homo sapiens})
Ensembl transcript IDs to gene IDs based on genome assembly
GRCH38 from Ensembl.
}
\details{
Variables:
\itemize{
\item enstxp
\item ensgene
}
}
\examples{
head(grch38_tx2gene)
}
\keyword{datasets}
