% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grcm38.R
\docType{data}
\name{grcm38}
\alias{grcm38}
\title{Mouse annotation GRCm38}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 50730 rows and 9 columns.}
\source{
\url{http://useast.ensembl.org/Mus_musculus}
}
\usage{
grcm38
}
\description{
Mouse annotation GRCm38 from Ensembl.
}
\examples{
head(grcm38)

}
\keyword{datasets}

