% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnor6.R
\docType{data}
\name{rnor6}
\alias{rnor6}
\title{Rat annotation Rnor_6.0}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 34200 rows and 9 columns.}
\source{
\url{http://useast.ensembl.org/Rattus_norvegicus}
}
\usage{
rnor6
}
\description{
Rat annotation Rnor_6.0 from Ensembl release 82.
}
\keyword{datasets}

