% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnor6.R
\docType{data}
\name{rnor6}
\alias{rnor6}
\title{Rat annotation data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 31048 rows and 9 columns.
}
\source{
\url{http://ensembl.org/rattus_norvegicus}
}
\usage{
rnor6
}
\description{
Rat (\emph{Rattus norvegicus}) annotations based on
genome assembly RNOR6 from Ensembl.
}
\details{
Variables:
\itemize{
\item ensgene
\item entrez
\item symbol
\item chr
\item start
\item end
\item strand
\item biotype
\item description
}
}
\examples{
head(rnor6)
}
\keyword{datasets}
