% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnor6_tx2gene.R
\docType{data}
\name{rnor6_tx2gene}
\alias{rnor6_tx2gene}
\title{Rat transcripts to genes}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 41078 rows and 2 columns.}
\source{
\url{http://ensembl.org/rattus_norvegicus}
}
\usage{
rnor6_tx2gene
}
\description{
Lookup table for converting Rat (\emph{Rattus norvegicus})
Ensembl transcript IDs to gene IDs based on genome assembly
RNOR6 from Ensembl.
}
\details{
Variables:
\itemize{
\item enstxp
\item ensgene
}
}
\examples{
head(rnor6_tx2gene)
}
\keyword{datasets}
