% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbcel235.R
\docType{data}
\name{wbcel235}
\alias{wbcel235}
\title{Worm annotation WBcel235}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 47065 rows and 9 columns.}
\source{
\url{http://useast.ensembl.org/Caenorhabditis_elegans}
}
\usage{
wbcel235
}
\description{
Worm annotation WBcel235 from Ensembl.
}
\examples{
head(wbcel235)

}
\keyword{datasets}

