% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbcel235_tx2gene.R
\docType{data}
\name{wbcel235_tx2gene}
\alias{wbcel235_tx2gene}
\title{Roundworm transcripts to genes}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 58941 rows and 2 columns.}
\source{
\url{http://ensembl.org/caenorhabditis_elegans}
}
\usage{
wbcel235_tx2gene
}
\description{
Lookup table for converting Roundworm (\emph{Caenorhabditis elegans})
Ensembl transcript IDs to gene IDs based on genome assembly
WBCEL235 from Ensembl.
}
\details{
Variables:
\itemize{
\item enstxp
\item ensgene
}
}
\examples{
head(wbcel235_tx2gene)
}
\keyword{datasets}
